+++ Sprint Sprint/Sprint Sprint/Framework/ModConfig.cs
@@ -11,6 +11,8 @@ namespace Sprint_Sprint.Framework
         public bool HoldToSprint { get; set; } = true;
         /// <summary> The sprinting speed </summary>
         public int SprintSpeed { get; set; } = 8;
+        /// <summary> The default horse speed, since sprinting while on mount is disabled. </summary>
+        public int HorseSpeed { get; set; } = 5;
         /// <summary> Modify stamina draining settings </summary>
         public StaminaDrainConfig StaminaDrain { get; set; } = new StaminaDrainConfig();
         /// <summary> Disable sprinting when player is too tired </summary>