+++ Sprint Sprint/docs/changelog.md
@@ -1,6 +1,10 @@
 # Sprint Sprint changelog
 [← README.md](../README.md)
 
+### 2.0.1
+- Fixed manifest file for those who use the Vortex mod manager
+- Added a new configuration option for Horse Speed, since the default horse speed is being overwritten and sprinting with horse is disabled
+
 ### 2.0.0
 - Initial release :D
 - Fixed the Sprint Sprint Sprint bugs